Refactored MAKEFILE
